+++ core/src/test/java/org/apache/pekko/persistence/jdbc/state/JavadslSnippets.java
