+++ core/src/test/scala/org/apache/pekko/persistence/jdbc/SharedActorSystemTestSpec.scala
@@ -6,13 +6,13 @@
 package org.apache.pekko.persistence.jdbc
 
 import org.apache.pekko.actor.ActorSystem
-import org.apache.pekko.persistence.jdbc.config.{ JournalConfig, ReadJournalConfig }
-import org.apache.pekko.persistence.jdbc.query.javadsl.JdbcReadJournal
-import org.apache.pekko.persistence.jdbc.util.DropCreate
-import org.apache.pekko.persistence.jdbc.db.SlickExtension
 import org.apache.pekko.serialization.SerializationExtension
 import org.apache.pekko.util.Timeout
-import com.typesafe.config.{ Config, ConfigFactory, ConfigValue }
+import com.typesafe.config.{Config, ConfigFactory, ConfigValue}
+import org.apache.pekko.persistence.jdbc.config.{JournalConfig, ReadJournalConfig}
+import org.apache.pekko.persistence.jdbc.db.SlickExtension
+import org.apache.pekko.persistence.jdbc.query.scaladsl.JdbcReadJournal
+import org.apache.pekko.persistence.jdbc.util.DropCreate
 import org.scalatest.BeforeAndAfterAll
 
 import scala.concurrent.ExecutionContext
