+++ core/src/test/scala/org/apache/pekko/persistence/jdbc/SimpleSpec.scala
@@ -5,11 +5,11 @@
 
 package org.apache.pekko.persistence.jdbc
 
-import org.apache.pekko.actor.{ ActorRef, ActorSystem }
+import org.apache.pekko.actor.{ActorRef, ActorSystem}
 import org.apache.pekko.persistence.jdbc.util.ClasspathResources
 import org.apache.pekko.testkit.TestProbe
 import org.scalatest._
-import org.scalatest.concurrent.{ Eventually, ScalaFutures }
+import org.scalatest.concurrent.{Eventually, ScalaFutures}
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
