+++ core/src/test/scala/org/apache/pekko/persistence/jdbc/SingleActorSystemPerTestSpec.scala
@@ -6,12 +6,12 @@
 package org.apache.pekko.persistence.jdbc
 
 import org.apache.pekko.actor.ActorSystem
-import org.apache.pekko.persistence.jdbc.config.{ JournalConfig, ReadJournalConfig, SlickConfiguration }
-import org.apache.pekko.persistence.jdbc.query.javadsl.JdbcReadJournal
-import org.apache.pekko.persistence.jdbc.util.DropCreate
-import org.apache.pekko.persistence.jdbc.db.SlickDatabase
 import org.apache.pekko.util.Timeout
-import com.typesafe.config.{ Config, ConfigFactory, ConfigValue }
+import com.typesafe.config.{Config, ConfigFactory, ConfigValue}
+import org.apache.pekko.persistence.jdbc.config.{JournalConfig, ReadJournalConfig, SlickConfiguration}
+import org.apache.pekko.persistence.jdbc.db.SlickDatabase
+import org.apache.pekko.persistence.jdbc.query.scaladsl.JdbcReadJournal
+import org.apache.pekko.persistence.jdbc.util.DropCreate
 import org.scalatest.BeforeAndAfterEach
 import slick.jdbc.JdbcBackend.Database
