+++ core/src/test/scala/org/apache/pekko/persistence/jdbc/configuration/ConfigOpsTest.scala
@@ -5,10 +5,9 @@
 
 package org.apache.pekko.persistence.jdbc.configuration
 
-import org.apache.pekko.persistence.jdbc.SimpleSpec
-import org.apache.pekko.persistence.jdbc.util.ConfigOps
-import ConfigOps._
 import com.typesafe.config.ConfigFactory
+import org.apache.pekko.persistence.jdbc.SimpleSpec
+import org.apache.pekko.persistence.jdbc.util.ConfigOps.ConfigOperations
 
 class ConfigOpsTest extends SimpleSpec {
   it should "parse field values to Options" in {
