+++ core/src/test/scala/org/apache/pekko/persistence/jdbc/configuration/JNDIConfigTest.scala
@@ -5,10 +5,10 @@
 
 package org.apache.pekko.persistence.jdbc.configuration
 
+import com.typesafe.config.ConfigFactory
 import org.apache.pekko.actor.ActorSystem
 import org.apache.pekko.persistence.jdbc.SimpleSpec
 import org.apache.pekko.persistence.jdbc.db.SlickExtension
-import com.typesafe.config.ConfigFactory
 
 class JNDIConfigTest extends SimpleSpec {
   "JNDI config" should "read the config and throw NoInitialContextException in case the JNDI resource is not available" in {
