+++ core/src/test/scala/org/apache/pekko/persistence/jdbc/configuration/PekkoPersistenceConfigTest.scala
@@ -5,14 +5,14 @@
 
 package org.apache.pekko.persistence.jdbc.configuration
 
-import org.apache.pekko.persistence.jdbc.config._
-import com.typesafe.config.{ Config, ConfigFactory }
+import com.typesafe.config.{Config, ConfigFactory}
+import org.apache.pekko.persistence.jdbc.config.{JournalConfig, ReadJournalConfig, SlickConfiguration, SnapshotConfig}
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 
 import scala.concurrent.duration._
 
-class AkkaPersistenceConfigTest extends AnyFlatSpec with Matchers {
+class PekkoPersistenceConfigTest extends AnyFlatSpec with Matchers {
   private val referenceConfig: Config = ConfigFactory.load("reference")
 
   val config: Config = ConfigFactory
