+++ core/src/test/scala/org/apache/pekko/persistence/jdbc/journal/dao/TagsSerializationTest.scala
@@ -3,9 +3,10 @@
  * Copyright (C) 2019 - 2021 Lightbend Inc. <https://www.lightbend.com>
  */
 
-package org.apache.pekko.persistence.jdbc.journal.dao.legacy
+package org.apache.pekko.persistence.jdbc.journal.dao
 
 import org.apache.pekko.persistence.jdbc.SharedActorSystemTestSpec
+import org.apache.pekko.persistence.jdbc.journal.dao.legacy.{decodeTags, encodeTags}
 
 class TagsSerializationTest extends SharedActorSystemTestSpec {
   "Encode" should "no tags" in {
