+++ core/src/test/scala/org/apache/pekko/persistence/jdbc/query/HardDeleteQueryTest.scala
