+++ core/src/test/scala/org/apache/pekko/persistence/jdbc/state/ScaladslSnippets.scala
